requirements.txt
SpeechRecognition==3.10.0
pyaudio==0.2.11
edge-tts==6.1.7
pygame==2.5.2
anthropic==0.7.8
openai==1.3.5
google-generativeai==0.3.1
sentence-transformers==2.2.2
numpy==1.24.3
faiss-cpu==1.7.4
pystray==0.19.4
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
